update: rename admin to admin_gui
