+++ src/schoolmanagementsystem/admin_gui/AdminWindow.form
