+++ src/schoolmanagementsystem/admin_gui/editStudentForm.form
