+++ src/schoolmanagementsystem/admin_gui/editStudentForm.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package schoolmanagementsystem.admin;
+package schoolmanagementsystem.admin_gui;
 
 /**
  *
